requirements.txt
python-telegram-bot==13.2
PyMySQL==1.0.2
Pillow==8.1.0
